requirements.txt
python-docx>=1.1.0
requests>=2.31.0

